Move the tensor shape utils
