--- sparse_autoencoder/autoencoder/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Sparse autoencoder model & components utils."""
