+++ sparse_autoencoder/utils/tensor_shape.py
